# [RWRoute] Further divide LOCAL nodes into EAST/WEST for UltraScale(+)