--- a/src/com/xilinx/rapidwright/rwroute/PartialRouter.java
+++ b/src/com/xilinx/rapidwright/rwroute/PartialRouter.java
@@ -256,7 +256,7 @@ public class PartialRouter extends RWRoute {
             preservedNet = routingGraph.getPreservedNet(sinkRnode);
             if (preservedNet != null && preservedNet != net) {
                 unpreserveNets.add(preservedNet);
-                assert(sinkRnode.getType() == RouteNodeType.EXCLUSIVE_SINK);
+                assert(sinkRnode.getType().isExclusiveSink());
             }
         }
 
@@ -599,7 +599,7 @@ public class PartialRouter extends RWRoute {
                 RouteNode sourceRnode = connection.getSourceRnode();
                 RouteNode sinkRnode = connection.getSinkRnode();
                 assert(sourceRnode.getType() == RouteNodeType.EXCLUSIVE_SOURCE);
-                assert(sinkRnode.getType() == RouteNodeType.EXCLUSIVE_SINK);
+                assert(sinkRnode.getType().isExclusiveSink());
 
                 // Even though this connection is not expected to have any routing yet,
                 // perform a rip up anyway in order to release any exclusive sinks