--- a/src/com/xilinx/rapidwright/rwroute/RouteNode.java
+++ b/src/com/xilinx/rapidwright/rwroute/RouteNode.java
@@ -123,6 +123,9 @@ public class RouteNode extends Node implements Comparable<RouteNode> {
                 assert(length == 0 ||
                         (length <= 3 && series == Series.Versal));
                 break;
+            case EXCLUSIVE_SINK:
+                assert(length == 0);
+                break;
             case EXCLUSIVE_SINK_EAST:
             case EXCLUSIVE_SINK_WEST:
                 assert(length == 0 ||