At least on UltraScale/UltraScale+, it looks like that many local nodes (`BOUNCE`/`BYPASS`/`INODE`/`INT_NODE_IMUX`) can be divided into east and western sides of an `INT` tile. Once you fall into one side or the other, you end up staying on that side forever more. This conclusion was reached by examining the now-expanded `TestNode.testNodeReachabilityUltraScale()` test.

Thus, during routing to `IMUX` sinks, it's important to choose the correct side otherwise a lot of effort will be wasted.

Building on top of https://github.com/Xilinx/RapidWright/pull/1095, in this PR certain nodes will have a new `RouteNodeType.LOCAL_{EAST,WEST}` type, along with certain `IMUX` sinks being assigned `RouteNodeType.EXCLUSIVE_SINK_{EAST,WEST}`.

On the UltraScale-s, `CTRL` sinks and `INT_NODE_GLOBAL` nodes can be reached from and reach both sides, respectively.

This optimization is not yet applied for Versal which currently assumes that all local nodes and sinks can reach both sides.