@eddieh-xlnx 2024-11-12
New results, again showing just the RWRoute wall time and the number of nodes it popped, for the FPGA24 Routing Contest benchmarks. Link to the before state: [PR#1100](#1100)
```
Before (PR#1100 @ abd15c7)
+--------------------------------+------+---------------+--------------+
| Benchmark                      | Pass | RWRoute (sec) | Nodes Popped |
+--------------------------------+------+---------------+--------------+
| logicnets_jscl_rwroute         | True |         43.83 |     67237191 |
| boom_med_pb_rwroute            | True |        234.27 |    263113332 |
| vtr_mcml_rwroute               | True |        238.19 |    344639097 |
| rosetta_fd_rwroute             | True |        153.83 |    206900150 |
| corundum_25g_rwroute           | True |        223.73 |    303466486 |
| finn_radioml_rwroute           | True |        101.62 |    138145459 |
| vtr_lu64peeng_rwroute          | True |        205.92 |    261583295 |
| corescore_500_rwroute          | True |        140.94 |    154687073 |
| corescore_500_pb_rwroute       | True |        258.28 |    287080189 |
| mlcad_d181_lefttwo3rds_rwroute | True |       1884.60 |   1689747584 |
| koios_dla_like_large_rwroute   | True |        318.21 |    352989376 |
| boom_soc_rwroute               | True |       1248.03 |   1589783049 |
| ispd16_example2_rwroute        | True |        507.40 |    601670342 |
+--------------------------------+------+---------------+--------------+

This PR @ 435efa8
+--------------------------------+------+---------------+--------------+
| Benchmark                      | Pass | RWRoute (sec) | Nodes Popped |
+--------------------------------+------+---------------+--------------+
| logicnets_jscl_rwroute         | True |         40.65 |     55611142 |
| boom_med_pb_rwroute            | True |        234.61 |    250431531 |
| vtr_mcml_rwroute               | True |        232.41 |    319777647 |
| rosetta_fd_rwroute             | True |        140.75 |    189200371 |
| corundum_25g_rwroute           | True |        212.33 |    267832074 |
| finn_radioml_rwroute           | True |         93.85 |    115969076 |
| vtr_lu64peeng_rwroute          | True |        190.44 |    224923115 |
| corescore_500_rwroute          | True |        114.45 |    122451976 |
| corescore_500_pb_rwroute       | True |        240.69 |    244333747 |
| mlcad_d181_lefttwo3rds_rwroute | True |       1972.12 |   1686723818 |
| koios_dla_like_large_rwroute   | True |        284.35 |    299644675 |
| boom_soc_rwroute               | True |       1196.53 |   1449104918 |
| ispd16_example2_rwroute        | True |        456.29 |    523581515 |
+--------------------------------+------+---------------+--------------+
```

In total, a 2.7% improvement in wall time but a more significant 8.2% reduction in number of nodes popped. The only noticeable regression appears to be on `mlcad_d181_lefttwo3rds` (4.6% increase in runtime, roughly same number of pops). Roughly the same number of routing iterations (80/79) too.